events that have been code deleted do not raise an error on load (#18)
Currently code deleting an event requires us to destroy any persisted
events in the database as well.
This is because the rails STI implementation raise a `SubclassNotFound`
exception if it is unable to find the event class.

lines -->
Override the exception behaviour and dynamically generate a no-op class
so the persisted event is effectively ignored.

<!--
Consider adding the following sections:

reviewers ]
ticket, e.g. SECURITY-123-human-readable-thing ]
-->