update timestamps for deleted events
the orignal event that was deleted would have set the `updated_at` and `created_at` timestamps on the projection.

This leads to drift in the projection from the events and would require a reprojection for every instance of the model.

To get around requiring this we will continue to update the timestamps for deleted events.