Guaranteed synchronous reactor order execution (#64)
#### Why <!-- A short description of why this change is required -->
Reactors were stored in a `Set` upon registration. Although ruby `Set`s
preserve order, this is not guaranteed in the future and depends on the
internal implementation of a `Set`. If we want to preserve order we
should be using an array.

Preserving order lets us guarantee the order or execution of synchronous
reactors (the order in which they were registered).


#### What changed <!-- Summary of changes when modifying hundreds of
lines -->
* Use an array to store reactors
* Update shared examples to accept either a single reactor or multiple
reactors